+++ src/topBar/dropdownButton/DropdownButton.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import cc from 'classcat'
 
-import { color } from '../_utils/branding'
-import { A11yProps, pickA11yProps } from '../_utils/interfaces'
-import { ChevronIcon } from '../icon/chevronIcon'
+import { color } from '../../_utils/branding'
+import { A11yProps, pickA11yProps } from '../../_utils/interfaces'
+import { ChevronIcon } from '../../icon/chevronIcon'
 import { StyledDropdownButton } from './DropdownButton.style'
 
 export type DropdownButtonProps = A11yProps &