+++ src/topBar/index.tsx
@@ -1,1 +1,4 @@
 export { TopBar, TopBarProps } from './TopBar'
+export { Drawer, DrawerProps } from './drawer'
+export { DropdownButton, DropdownButtonProps } from './dropdownButton'
+export { Menu, MenuProps } from './menu'