+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- **[BREAKING CHANGE]** Move `Drawer`, `DropdownButton`, `HamburgerButton` and `Menu` to `TopBar`
 - **[BREAKING CHANGE]** Move `Proximity` to `newItinerary`
 - **[BREAKING CHANGE]** Removed `small` prop on `TimePicker`
 - **[BREAKING CHANGE]** `Stars` and `Rating` components are now private